Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-backgrounds/animations/border-image-slice-composition-expected.txt


FAIL Compositing: property <border-image-slice> underlying [1 2 3 4] from add [1 2 3 4] to add [101 102 103 104] at (-0.25) should be [0] assert_equals: expected "0 " but got "- 24 - 23 - 22 - 21 "
FAIL Compositing: property <border-image-slice> underlying [1 2 3 4] from add [1 2 3 4] to add [101 102 103 104] at (0) should be [2 4 6 8] assert_equals: expected "2 4 6 8 " but got "1 2 3 4 "
FAIL Compositing: property <border-image-slice> underlying [1 2 3 4] from add [1 2 3 4] to add [101 102 103 104] at (0.25) should be [27 29 31 33] assert_equals: expected "27 29 31 33 " but got "26 27 28 29 "
FAIL Compositing: property <border-image-slice> underlying [1 2 3 4] from add [1 2 3 4] to add [101 102 103 104] at (0.5) should be [52 54 56 58] assert_equals: expected "52 54 56 58 " but got "51 52 53 54 "
FAIL Compositing: property <border-image-slice> underlying [1 2 3 4] from add [1 2 3 4] to add [101 102 103 104] at (0.75) should be [77 79 81 83] assert_equals: expected "77 79 81 83 " but got "76 77 78 79 "
FAIL Compositing: property <border-image-slice> underlying [1 2 3 4] from add [1 2 3 4] to add [101 102 103 104] at (1) should be [102 104 106 108] assert_equals: expected "102 104 106 108 " but got "101 102 103 104 "
FAIL Compositing: property <border-image-slice> underlying [1 2 3 4] from add [1 2 3 4] to add [101 102 103 104] at (1.25) should be [127 129 131 133] assert_equals: expected "127 129 131 133 " but got "126 127 128 129 "
FAIL Compositing: property <border-image-slice> underlying [100 200 300 400 fill] from add [100 fill] to add [200 300 500 fill] at (-0.25) should be [175 250 300 450 fill] assert_equals: expected "175 250 300 450 fill " but got "75 50 0 fill "
FAIL Compositing: property <border-image-slice> underlying [100 200 300 400 fill] from add [100 fill] to add [200 300 500 fill] at (0) should be [200 300 400 500 fill] assert_equals: expected "200 300 400 500 fill " but got "100 fill "
FAIL Compositing: property <border-image-slice> underlying [100 200 300 400 fill] from add [100 fill] to add [200 300 500 fill] at (0.25) should be [225 350 500 550 fill] assert_equals: expected "225 350 500 550 fill " but got "125 150 200 fill "
FAIL Compositing: property <border-image-slice> underlying [100 200 300 400 fill] from add [100 fill] to add [200 300 500 fill] at (0.5) should be [250 400 600 600 fill] assert_equals: expected "250 400 600 600 fill " but got "150 200 300 fill "
FAIL Compositing: property <border-image-slice> underlying [100 200 300 400 fill] from add [100 fill] to add [200 300 500 fill] at (0.75) should be [275 450 700 650 fill] assert_equals: expected "275 450 700 650 fill " but got "175 250 400 fill "
FAIL Compositing: property <border-image-slice> underlying [100 200 300 400 fill] from add [100 fill] to add [200 300 500 fill] at (1) should be [300 500 800 700 fill] assert_equals: expected "300 500 800 700 fill " but got "200 300 500 fill "
FAIL Compositing: property <border-image-slice> underlying [100 200 300 400 fill] from add [100 fill] to add [200 300 500 fill] at (1.25) should be [325 550 900 750 fill] assert_equals: expected "325 550 900 750 fill " but got "225 350 600 fill "
FAIL Compositing: property <border-image-slice> underlying [1 2 3% 4%] from add [1 2 3% 4%] to add [101 102 103% 104%] at (-0.25) should be [0 0 0% 0%] assert_equals: expected "0 0 0 % 0 % " but got "- 24 - 23 - 22 % - 21 % "
FAIL Compositing: property <border-image-slice> underlying [1 2 3% 4%] from add [1 2 3% 4%] to add [101 102 103% 104%] at (0) should be [2 4 6% 8%] assert_equals: expected "2 4 6 % 8 % " but got "1 2 3 % 4 % "
FAIL Compositing: property <border-image-slice> underlying [1 2 3% 4%] from add [1 2 3% 4%] to add [101 102 103% 104%] at (0.25) should be [27 29 31% 33%] assert_equals: expected "27 29 31 % 33 % " but got "26 27 28 % 29 % "
FAIL Compositing: property <border-image-slice> underlying [1 2 3% 4%] from add [1 2 3% 4%] to add [101 102 103% 104%] at (0.5) should be [52 54 56% 58%] assert_equals: expected "52 54 56 % 58 % " but got "51 52 53 % 54 % "
FAIL Compositing: property <border-image-slice> underlying [1 2 3% 4%] from add [1 2 3% 4%] to add [101 102 103% 104%] at (0.75) should be [77 79 81% 83%] assert_equals: expected "77 79 81 % 83 % " but got "76 77 78 % 79 % "
FAIL Compositing: property <border-image-slice> underlying [1 2 3% 4%] from add [1 2 3% 4%] to add [101 102 103% 104%] at (1) should be [102 104 106% 108%] assert_equals: expected "102 104 106 % 108 % " but got "101 102 103 % 104 % "
FAIL Compositing: property <border-image-slice> underlying [1 2 3% 4%] from add [1 2 3% 4%] to add [101 102 103% 104%] at (1.25) should be [127 129 131% 133%] assert_equals: expected "127 129 131 % 133 % " but got "126 127 128 % 129 % "
FAIL Compositing: property <border-image-slice> underlying [10% 20%] from add [190% 180% 290% 280%] to add [90% 80%] at (-0.25) should be [225% 225% 350% 350%] assert_equals: expected "225 % 225 % 350 % 350 % " but got "215 % 205 % 340 % 330 % "
FAIL Compositing: property <border-image-slice> underlying [10% 20%] from add [190% 180% 290% 280%] to add [90% 80%] at (0) should be [200% 200% 300% 300%] assert_equals: expected "200 % 200 % 300 % 300 % " but got "190 % 180 % 290 % 280 % "
FAIL Compositing: property <border-image-slice> underlying [10% 20%] from add [190% 180% 290% 280%] to add [90% 80%] at (0.25) should be [175% 175% 250% 250%] assert_equals: expected "175 % 175 % 250 % 250 % " but got "165 % 155 % 240 % 230 % "
FAIL Compositing: property <border-image-slice> underlying [10% 20%] from add [190% 180% 290% 280%] to add [90% 80%] at (0.5) should be [150% 150% 200% 200%] assert_equals: expected "150 % 150 % 200 % 200 % " but got "140 % 130 % 190 % 180 % "
FAIL Compositing: property <border-image-slice> underlying [10% 20%] from add [190% 180% 290% 280%] to add [90% 80%] at (0.75) should be [125% 125% 150% 150%] assert_equals: expected "125 % 125 % 150 % 150 % " but got "115 % 105 % 140 % 130 % "
FAIL Compositing: property <border-image-slice> underlying [10% 20%] from add [190% 180% 290% 280%] to add [90% 80%] at (1) should be [100%] assert_equals: expected "100 % " but got "90 % 80 % "
FAIL Compositing: property <border-image-slice> underlying [10% 20%] from add [190% 180% 290% 280%] to add [90% 80%] at (1.25) should be [75% 75% 50% 50%] assert_equals: expected "75 % 75 % 50 % 50 % " but got "65 % 55 % 40 % 30 % "
FAIL Compositing: property <border-image-slice> underlying [10 20%] from replace [100 100%] to add [190 180%] at (-0.25) should be [75 75%] assert_equals: expected "75 75 % " but got "77.5 80 % "
PASS Compositing: property <border-image-slice> underlying [10 20%] from replace [100 100%] to add [190 180%] at (0) should be [100 100%]
FAIL Compositing: property <border-image-slice> underlying [10 20%] from replace [100 100%] to add [190 180%] at (0.25) should be [125 125%] assert_equals: expected "125 125 % " but got "122.5 120 % "
FAIL Compositing: property <border-image-slice> underlying [10 20%] from replace [100 100%] to add [190 180%] at (0.5) should be [150 150%] assert_equals: expected "150 150 % " but got "145 140 % "
FAIL Compositing: property <border-image-slice> underlying [10 20%] from replace [100 100%] to add [190 180%] at (0.75) should be [175 175%] assert_equals: expected "175 175 % " but got "167.5 160 % "
FAIL Compositing: property <border-image-slice> underlying [10 20%] from replace [100 100%] to add [190 180%] at (1) should be [200 200%] assert_equals: expected "200 200 % " but got "190 180 % "
FAIL Compositing: property <border-image-slice> underlying [10 20%] from replace [100 100%] to add [190 180%] at (1.25) should be [225 225%] assert_equals: expected "225 225 % " but got "212.5 200 % "
FAIL Compositing: property <border-image-slice> underlying [10% 20] from add [90% 80] to replace [0% 0 0% 0] at (-0.25) should be [125% 125] assert_equals: expected "125 % 125 " but got "112.5 % 100 "
FAIL Compositing: property <border-image-slice> underlying [10% 20] from add [90% 80] to replace [0% 0 0% 0] at (0) should be [100% 100] assert_equals: expected "100 % 100 " but got "90 % 80 "
FAIL Compositing: property <border-image-slice> underlying [10% 20] from add [90% 80] to replace [0% 0 0% 0] at (0.25) should be [75% 75] assert_equals: expected "75 % 75 " but got "67.5 % 60 "
FAIL Compositing: property <border-image-slice> underlying [10% 20] from add [90% 80] to replace [0% 0 0% 0] at (0.5) should be [50% 50] assert_equals: expected "50 % 50 " but got "45 % 40 "
FAIL Compositing: property <border-image-slice> underlying [10% 20] from add [90% 80] to replace [0% 0 0% 0] at (0.75) should be [25% 25] assert_equals: expected "25 % 25 " but got "22.5 % 20 "
PASS Compositing: property <border-image-slice> underlying [10% 20] from add [90% 80] to replace [0% 0 0% 0] at (1) should be [0% 0]
FAIL Compositing: property <border-image-slice> underlying [10% 20] from add [90% 80] to replace [0% 0 0% 0] at (1.25) should be [0% 0] assert_equals: expected "0 % 0 " but got "- 22.5 % - 20 "
FAIL Compositing: property <border-image-slice> underlying [10 20] from add [100% 150%] to add [200% 250% fill] at (-0.25) should be [100% 150%] assert_equals: expected "100 % 150 % " but got "75 % 125 % "
PASS Compositing: property <border-image-slice> underlying [10 20] from add [100% 150%] to add [200% 250% fill] at (0) should be [100% 150%]
FAIL Compositing: property <border-image-slice> underlying [10 20] from add [100% 150%] to add [200% 250% fill] at (0.25) should be [100% 150%] assert_equals: expected "100 % 150 % " but got "125 % 175 % "
FAIL Compositing: property <border-image-slice> underlying [10 20] from add [100% 150%] to add [200% 250% fill] at (0.5) should be [200% 250% fill] assert_equals: expected "200 % 250 % fill " but got "150 % 200 % "
FAIL Compositing: property <border-image-slice> underlying [10 20] from add [100% 150%] to add [200% 250% fill] at (0.75) should be [200% 250% fill] assert_equals: expected "200 % 250 % fill " but got "175 % 225 % "
FAIL Compositing: property <border-image-slice> underlying [10 20] from add [100% 150%] to add [200% 250% fill] at (1) should be [200% 250% fill] assert_equals: expected "200 % 250 % fill " but got "200 % 250 % "
FAIL Compositing: property <border-image-slice> underlying [10 20] from add [100% 150%] to add [200% 250% fill] at (1.25) should be [200% 250% fill] assert_equals: expected "200 % 250 % fill " but got "225 % 275 % "
PASS Compositing: property <border-image-slice> underlying [10 20] from add [100 150%] to add [200% 250] at (-0.25) should be [100 150%]
PASS Compositing: property <border-image-slice> underlying [10 20] from add [100 150%] to add [200% 250] at (0) should be [100 150%]
FAIL Compositing: property <border-image-slice> underlying [10 20] from add [100 150%] to add [200% 250] at (0.25) should be [100 150%] assert_equals: expected "100 150 % " but got "0 "
FAIL Compositing: property <border-image-slice> underlying [10 20] from add [100 150%] to add [200% 250] at (0.5) should be [200% 250] assert_equals: expected "200 % 250 " but got "0 "
FAIL Compositing: property <border-image-slice> underlying [10 20] from add [100 150%] to add [200% 250] at (0.75) should be [200% 250] assert_equals: expected "200 % 250 " but got "0 "
PASS Compositing: property <border-image-slice> underlying [10 20] from add [100 150%] to add [200% 250] at (1) should be [200% 250]
PASS Compositing: property <border-image-slice> underlying [10 20] from add [100 150%] to add [200% 250] at (1.25) should be [200% 250]
